remove regex and move it test to streaming
